requirements.txt
streamlit
pymupdf
opencv-python
Pillow
numpy
